refactor(pull-request): final cleanup for pr refactor
