+++ src/cmds/pull-request/info.ts
@@ -3,12 +3,29 @@
  * (see file: README.md)
  * SPDX-License-Identifier: BSD-3-Clause
  */
-import { get } from './index'
+import { printPullInfo } from './index'
+import * as logger from '../../logger'
 
 export async function infoHandler(options) {
     try {
-        await get(options, options.user, options.repo, options.number)
+        await info(options, options.user, options.repo, options.number)
     } catch (err) {
         throw new Error(`Can't get pull requests.\n${err}`)
     }
+}
+
+export async function info(options, user, repo, number) {
+    const payload = {
+        repo,
+        pull_number: number,
+        owner: user,
+    }
+
+    try {
+        var { data: pull } = await options.GitHub.pulls.get(payload)
+    } catch (err) {
+        logger.warn(`Can't get pull request ${user}/${repo}/${number}`)
+    }
+
+    printPullInfo(options, pull)
 }
